Creates the iterator_reducation.cpp file to track the refactoring that has happened so far with iterator.cpp.
Removing this file from the src/ directory allows CMake and gcc to compile TDMS as usual.
The purpose of this file is to have a live-progress report of the refactor - replacing code with refactored methods as they are pulled out.

The main loop will not be touched, just moved (as it appears) into a "function method" in it's own file.
The purpose of the classes currently is simply to make the setup and tear-down of what is currently execute_simulation more readable and manageable!