Introduce a versioning system. (#203)
* Resolves #201.

Add a mechanism for CMake to read a git hash and construct a version. A
tag of semver format on main is taken as an official version, otherwise
construct a placeholder <branchname>_<tag or short format commit hash>.

* Mosè's comment from PR #203.

If the user grabs a snapshot tarball then there is no git repo so fall
emit a friendly warning message and fall back to the default release
name defined in globals.h.

* Test the --version option.

* Apply suggestions from code review

Co-authored-by: Will Graham <32364977+willGraham01@users.noreply.github.com>

---------

Co-authored-by: Will Graham <32364977+willGraham01@users.noreply.github.com>